refactor: restruct sidebar nav componenet
