+++ apps/web/src/widgets/navigation-layout/ui/index.ts
@@ -1,4 +1,4 @@
 export { NavigationLayout } from './NavigationLayout';
-export { SidebarNav } from './SideBarNav';
+export { SidebarNav } from './sidebar';
 export { BottomNav } from './BottomNav';
 export { FooterProfile } from './FooterProfile';
