+++ apps/web/src/widgets/navigation-layout/ui/sidebar/SidebarHeader.tsx
@@ -1,0 +1,4 @@
+export function SideBarHeader() {
+  /* TODO: 이미지 로고 대체 예정 */
+  return <span className="px-2.5 font-bold text-blue-500">크자회 Logo</span>;
+}
