+++ apps/web/src/widgets/navigation-layout/ui/sidebar/index.ts
@@ -1,0 +1,1 @@
+export * from './SideBarNav';
